contributors/ttwf/rickhurst/submitted/2d-rotate-js.html

<!DOCTYPE html>
<html>
	<head>
		<title>JS test: Rotate via javascript must show the correct computed rotation</title>
		<link rel="author" title="Rick Hurst" href="http://mrkn.co/axegs">
		<link rel="help" href="http://www.w3.org/TR/css3-transforms/#transform-property">
		<meta name="flags" content="css3, rotate, svg">
		<meta name="assert" content="Asserting that you can rotate an element with JS and it show up in CSS computed values not as a matrix but as the rotation">
		<style>
			#box{
				margin-top:30px;
				display: block;
				width: 50px;
				height: 50px;
				background-color: green;
			}
		</style>
	</head>
    <body>
		<h1>Rotate via JS</h1>
        <div id="log"></div>
        <div id="box"></div>
		<script src="/resources/testharness.js"></script>
		<script src="/resources/testharnessreport.js"></script>
		<script>
			(function(){
				var box = document.getElementById("box"),
					assertion = "",
					expectation = "rotate(30deg)",
					should_string = "this should make a small green square rotated 30 degrees, and the browser should return the rotation as 30 degrees as the computed value NOT a matrix",
					extra_properties = null,
					computed_angle = 0,
					a_1 = "";
					a_2 = "",
					my_value = "",
					test_function = {};
	
				/*set it to rotate 30 degrees*/
				box.style.MozTransform = "rotate(30deg)";
				
				/*  All browsers return a matrix "currently" but I feel that
					it may be more useful to to return the value that the user set, 
					so they don't have to convert matrix to rotate values in their code.
					
					With something like the following:
					
				/**
                 * Matrix To Rotation String
                 *
                 * @param (String) matrix_string // matrix(0.866025, 0.5, -0.5, 0.866025, 0, 0)
                 * @return {Number} rotation in degrees 0.5
                 */                      
                 /*
                 function matrixToRotationString(matrix_string){
                     a_1 = matrix_string.replace("matrix(", "");
                     a_2 = a_1.substring(0, a_1.length-1);
                     my_value = a_2.split(",");
                     return "rotate(" + Math.round(Math.asin(parseFloat(my_value[1])) * (180/Math.PI)) + "deg)";    
                 }
                 
				 */
					
				
				assertion = window.getComputedStyle(box).getPropertyValue("-moz-transform");
			
				test_function = function(){
					assert_equals(assertion, expectation);
					//assert_equals(matrixToRotationString(assertion), expectation);	
				}		
				test(test_function, should_string);
				
			})();	
		</script>
    </body>
</html>
